feat: Implement `<i>` and `<span>` elements with comprehensive unit tests for behavior and rendering.
