+++ tests/Phrasing/SpanTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace UIAwesome\Html\Tests\Flow;
+namespace UIAwesome\Html\Tests\Phrasing;
 
 use PHPForge\Support\LineEndingNormalizer;
 use PHPUnit\Framework\Attributes\Group;
@@ -17,35 +17,35 @@ use UIAwesome\Html\Attribute\Values\{
     Translate,
 };
 use UIAwesome\Html\Core\Factory\SimpleFactory;
-use UIAwesome\Html\Flow\I;
+use UIAwesome\Html\Phrasing\Span;
 use UIAwesome\Html\Tests\Support\Stub\{DefaultProvider, DefaultThemeProvider};
 
 /**
- * Unit tests for {@see I} `<i>` behavior.
+ * Unit tests for {@see Span} `<span>` behavior.
  *
- * Verifies rendered output, attribute handling, configuration precedence, and content encoding for {@see I::tag()}.
+ * Verifies rendered output, attribute handling, configuration precedence, and content encoding for {@see Span::tag()}.
  *
  * Test coverage.
  * - Applies global `aria-*` and `data-*` attributes via helper methods.
  * - Applies global defaults and theme providers via {@see SimpleFactory} and provider stubs.
  * - Renders content and string casting for an inline element.
  *
- * {@see I} for implementation details.
+ * {@see Span} for implementation details.
  *
  * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
 #[Group('html')]
 #[Group('flow')]
-final class ITest extends TestCase
+final class SpanTest extends TestCase
 {
     public function testContentEncodesValues(): void
     {
-        $i = I::tag()->content('<value>');
+        $span = Span::tag()->content('<value>');
 
         self::assertSame(
             '&lt;value&gt;',
-            $i->getContent(),
+            $span->getContent(),
             "Failed asserting that 'content()' method encodes values correctly.",
         );
     }
@@ -54,7 +54,7 @@ final class ITest extends TestCase
     {
         self::assertSame(
             'default',
-            I::tag()->getAttribute('data-test', 'default'),
+            Span::tag()->getAttribute('data-test', 'default'),
             "Failed asserting that 'getAttribute()' returns the default value when missing.",
         );
     }
@@ -63,7 +63,7 @@ final class ITest extends TestCase
     {
         self::assertSame(
             ['data-test' => 'value'],
-            I::tag()->addAttribute('data-test', 'value')->getAttributes(),
+            Span::tag()->addAttribute('data-test', 'value')->getAttributes(),
             "Failed asserting that 'getAttributes()' returns the assigned attributes.",
         );
     }
@@ -72,10 +72,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i><value></i>
+            <span><value></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->html('<value>')->render(),
+                Span::tag()->html('<value>')->render(),
             ),
             "Failed asserting that element renders correctly with 'html()' method.",
         );
@@ -85,10 +85,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i accesskey="k"></i>
+            <span accesskey="k"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->accesskey('k')->render(),
+                Span::tag()->accesskey('k')->render(),
             ),
             "Failed asserting that element renders correctly with 'accesskey' attribute.",
         );
@@ -98,10 +98,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i aria-label="Idiomatic text"></i>
+            <span aria-label="Span content"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->addAriaAttribute('label', 'Idiomatic text')->render(),
+                Span::tag()->addAriaAttribute('label', 'Span content')->render(),
             ),
             "Failed asserting that element renders correctly with 'addAriaAttribute()' method.",
         );
@@ -111,10 +111,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i aria-hidden="true"></i>
+            <span aria-hidden="true"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->addAriaAttribute(Aria::HIDDEN, true)->render(),
+                Span::tag()->addAriaAttribute(Aria::HIDDEN, true)->render(),
             ),
             "Failed asserting that element renders correctly with 'addAriaAttribute()' method using enum.",
         );
@@ -124,10 +124,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i data-test="value"></i>
+            <span data-test="value"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->addAttribute('data-test', 'value')->render(),
+                Span::tag()->addAttribute('data-test', 'value')->render(),
             ),
             "Failed asserting that element renders correctly with 'addAttribute()' method.",
         );
@@ -137,10 +137,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i title="Idiomatic text"></i>
+            <span title="Span content"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->addAttribute(GlobalAttribute::TITLE, 'Idiomatic text')->render(),
+                Span::tag()->addAttribute(GlobalAttribute::TITLE, 'Span content')->render(),
             ),
             "Failed asserting that element renders correctly with 'addAttribute()' method using enum.",
         );
@@ -150,10 +150,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i data-value="value"></i>
+            <span data-value="value"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->addDataAttribute('value', 'value')->render(),
+                Span::tag()->addDataAttribute('value', 'value')->render(),
             ),
             "Failed asserting that element renders correctly with 'addDataAttribute()' method.",
         );
@@ -163,10 +163,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i data-value="value"></i>
+            <span data-value="value"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->addDataAttribute(Data::VALUE, 'value')->render(),
+                Span::tag()->addDataAttribute(Data::VALUE, 'value')->render(),
             ),
             "Failed asserting that element renders correctly with 'addDataAttribute()' method using enum.",
         );
@@ -176,10 +176,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i aria-controls="modal-1" aria-hidden="false" aria-label="Close"></i>
+            <span aria-controls="modal-1" aria-hidden="false" aria-label="Close"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()
+                Span::tag()
                     ->ariaAttributes(
                         [
                             'controls' => static fn(): string => 'modal-1',
@@ -197,10 +197,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i class="value"></i>
+            <span class="value"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->attributes(['class' => 'value'])->render(),
+                Span::tag()->attributes(['class' => 'value'])->render(),
             ),
             "Failed asserting that element renders correctly with 'attributes()' method.",
         );
@@ -210,10 +210,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i class="value"></i>
+            <span class="value"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->class('value')->render(),
+                Span::tag()->class('value')->render(),
             ),
             "Failed asserting that element renders correctly with 'class' attribute.",
         );
@@ -223,10 +223,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i>value</i>
+            <span>value</span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->content('value')->render(),
+                Span::tag()->content('value')->render(),
             ),
             'Failed asserting that element renders correctly with default values.',
         );
@@ -236,10 +236,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i data-value="value"></i>
+            <span data-value="value"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->dataAttributes(['value' => 'value'])->render(),
+                Span::tag()->dataAttributes(['value' => 'value'])->render(),
             ),
             "Failed asserting that element renders correctly with 'dataAttributes()' method.",
         );
@@ -249,10 +249,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i class="default-class"></i>
+            <span class="default-class"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag(['class' => 'default-class'])->render(),
+                Span::tag(['class' => 'default-class'])->render(),
             ),
             'Failed asserting that default configuration values are applied correctly.',
         );
@@ -262,10 +262,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i class="default-class" title="default-title"></i>
+            <span class="default-class" title="default-title"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->addDefaultProvider(DefaultProvider::class)->render(),
+                Span::tag()->addDefaultProvider(DefaultProvider::class)->render(),
             ),
             'Failed asserting that default provider is applied correctly.',
         );
@@ -275,10 +275,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i></i>
+            <span></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->render(),
+                Span::tag()->render(),
             ),
             'Failed asserting that element renders correctly with default values.',
         );
@@ -288,10 +288,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i dir="ltr"></i>
+            <span dir="ltr"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->dir('ltr')->render(),
+                Span::tag()->dir('ltr')->render(),
             ),
             "Failed asserting that element renders correctly with 'dir' attribute.",
         );
@@ -301,10 +301,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i dir="ltr"></i>
+            <span dir="ltr"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->dir(Direction::LTR)->render(),
+                Span::tag()->dir(Direction::LTR)->render(),
             ),
             "Failed asserting that element renders correctly with 'dir' attribute using enum.",
         );
@@ -312,29 +312,29 @@ final class ITest extends TestCase
 
     public function testRenderWithGlobalDefaultsAreApplied(): void
     {
-        SimpleFactory::setDefaults(I::class, ['class' => 'default-class']);
+        SimpleFactory::setDefaults(Span::class, ['class' => 'default-class']);
 
         self::assertSame(
             <<<HTML
-            <i class="default-class"></i>
+            <span class="default-class"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->render(),
+                Span::tag()->render(),
             ),
             'Failed asserting that global defaults are applied correctly.',
         );
 
-        SimpleFactory::setDefaults(I::class, []);
+        SimpleFactory::setDefaults(Span::class, []);
     }
 
     public function testRenderWithHidden(): void
     {
         self::assertSame(
             <<<HTML
-            <i hidden></i>
+            <span hidden></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->hidden(true)->render(),
+                Span::tag()->hidden(true)->render(),
             ),
             "Failed asserting that element renders correctly with 'hidden' attribute.",
         );
@@ -344,10 +344,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i id="test-id"></i>
+            <span id="test-id"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->id('test-id')->render(),
+                Span::tag()->id('test-id')->render(),
             ),
             "Failed asserting that element renders correctly with 'id' attribute.",
         );
@@ -357,10 +357,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i lang="es"></i>
+            <span lang="es"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->lang('es')->render(),
+                Span::tag()->lang('es')->render(),
             ),
             "Failed asserting that element renders correctly with 'lang' attribute.",
         );
@@ -370,10 +370,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i lang="es"></i>
+            <span lang="es"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->lang(Language::SPANISH)->render(),
+                Span::tag()->lang(Language::SPANISH)->render(),
             ),
             "Failed asserting that element renders correctly with 'lang' attribute using enum.",
         );
@@ -383,10 +383,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i></i>
+            <span></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()
+                Span::tag()
                     ->addAriaAttribute('label', 'Close')
                     ->removeAriaAttribute('label')
                     ->render(),
@@ -399,10 +399,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i></i>
+            <span></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()
+                Span::tag()
                     ->addAttribute('data-test', 'value')
                     ->removeAttribute('data-test')
                     ->render(),
@@ -415,10 +415,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i></i>
+            <span></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()
+                Span::tag()
                     ->addDataAttribute('value', 'test')
                     ->removeDataAttribute('value')
                     ->render(),
@@ -431,10 +431,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i role="button"></i>
+            <span role="button"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->role('button')->render(),
+                Span::tag()->role('button')->render(),
             ),
             "Failed asserting that element renders correctly with 'role' attribute.",
         );
@@ -444,10 +444,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i role="button"></i>
+            <span role="button"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->role(Role::BUTTON)->render(),
+                Span::tag()->role(Role::BUTTON)->render(),
             ),
             "Failed asserting that element renders correctly with 'role' attribute using enum.",
         );
@@ -457,10 +457,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i style='color: red;'></i>
+            <span style='color: red;'></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->style('color: red;')->render(),
+                Span::tag()->style('color: red;')->render(),
             ),
             "Failed asserting that element renders correctly with 'style' attribute.",
         );
@@ -470,10 +470,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i class="text-muted"></i>
+            <span class="text-muted"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->addThemeProvider('muted', DefaultThemeProvider::class)->render(),
+                Span::tag()->addThemeProvider('muted', DefaultThemeProvider::class)->render(),
             ),
             "Failed asserting that element renders correctly with 'addThemeProvider()' method.",
         );
@@ -483,10 +483,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i title="value"></i>
+            <span title="value"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->title('value')->render(),
+                Span::tag()->title('value')->render(),
             ),
             "Failed asserting that element renders correctly with 'title' attribute.",
         );
@@ -496,10 +496,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i></i>
+            <span></span>
             HTML,
             LineEndingNormalizer::normalize(
-                (string) I::tag(),
+                (string) Span::tag(),
             ),
             "Failed asserting that '__toString()' method renders correctly.",
         );
@@ -509,10 +509,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i translate="no"></i>
+            <span translate="no"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->translate(false)->render(),
+                Span::tag()->translate(false)->render(),
             ),
             "Failed asserting that element renders correctly with 'translate' attribute.",
         );
@@ -522,10 +522,10 @@ final class ITest extends TestCase
     {
         self::assertSame(
             <<<HTML
-            <i translate="no"></i>
+            <span translate="no"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag()->translate(Translate::NO)->render(),
+                Span::tag()->translate(Translate::NO)->render(),
             ),
             "Failed asserting that element renders correctly with 'translate' attribute using enum.",
         );
@@ -533,18 +533,18 @@ final class ITest extends TestCase
 
     public function testRenderWithUserOverridesGlobalDefaults(): void
     {
-        SimpleFactory::setDefaults(I::class, ['class' => 'from-global', 'id' => 'id-global']);
+        SimpleFactory::setDefaults(Span::class, ['class' => 'from-global', 'id' => 'id-global']);
 
         self::assertSame(
             <<<HTML
-            <i class="from-global" id="id-user"></i>
+            <span class="from-global" id="id-user"></span>
             HTML,
             LineEndingNormalizer::normalize(
-                I::tag(['id' => 'id-user'])->render(),
+                Span::tag(['id' => 'id-user'])->render(),
             ),
             'Failed asserting that user-defined attributes override global defaults correctly.',
         );
 
-        SimpleFactory::setDefaults(I::class, []);
+        SimpleFactory::setDefaults(Span::class, []);
     }
 }
